crud de Movies,genres, e cadastro e consulta de actors

+++ src/main/java/br/com/meli/movies1/repository/IMovieRepository.java
@@ -4,5 +4,5 @@ package br.com.meli.movies1.repository;
 import br.com.meli.movies1.model.Movies;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IMovieRepository extends JpaRepository<Movies, Long> {
+public interface IMovieRepository extends JpaRepository<Movies, Integer> {
 }
